Replaced default argument min_df from 20 to 0 on TFIDF (#213)
* Remove default min_df=20 argument from TFIDF

On sklearn min_df defaults to 0, it should default to 0 as well in fklearn. This argument is not specified in the docstring and was messing up with the performance of my sentiment classifier. I lost 30 points of recall because of this argument and had to spent hours figuring out that this was the problem. Could we change that  
argument to be consistent with sklearn?

* Update src/fklearn/training/classification.py

Co-authored-by: Hellen Lima <hellen.lima@nubank.com.br>

Co-authored-by: Hellen Lima <hellen.lima@nubank.com.br>